requirements.txt
Flask==2.2.2
numpy==1.23.4
SoundCard==0.4.2
soundfile==0.11.0
Flask-SocketIO==5.3.2
openai==0.25.0
google-cloud-speech==2.16.2
gevent==22.10.2
gevent-websocket==0.10.1
gunicorn==20.1.0
flask-sqlalchemy
flask_cors
psycopg2==2.9.5
SQLAlchemy==1.4.46
better-profanity==0.7.0
googletrans==3.1.0a0
pyaudio